+++ Mailer/extensions/mailer-module/src/Forms/JobFormFactory.php
@@ -7,6 +7,7 @@ use Nette\Application\UI\Form;
 use Nette\SmartObject;
 use Nette\Utils\ArrayHash;
 use Remp\MailerModule\Models\FormRenderer\MaterialRenderer;
+use Remp\MailerModule\Models\Job\JobSegmentsManager;
 use Remp\MailerModule\Models\Segment\Aggregator;
 use Remp\MailerModule\Repositories\BatchesRepository;
 use Remp\MailerModule\Repositories\JobsRepository;
@@ -61,13 +62,28 @@ class JobFormFactory
             Debugger::log($this->segmentAggregator->getErrors()[0], Debugger::WARNING);
         }
 
-        $form->addSelect('segment_code', 'Segment', $segments)
-            ->setPrompt('Select segment')
-            ->setRequired("Field 'Segment' is required.")
+        $jobSegmentsManager = new JobSegmentsManager($job);
+
+        $includeSegmentsDefault = array_map(static function (array $includeSegment) {
+            return $includeSegment['provider'] . '::' . $includeSegment['code'];
+        }, $jobSegmentsManager->getIncludeSegments());
+
+        $form->addMultiSelect('include_segment_codes', 'Include segments', $segments)
+            ->setRequired("You have to include at least one segment.")
             ->setHtmlAttribute('class', 'selectpicker')
             ->setHtmlAttribute('data-live-search', 'true')
             ->setHtmlAttribute('data-live-search-normalize', 'true')
-            ->setDefaultValue($job->segment_provider . '::' . $job->segment_code);
+            ->setDefaultValue($includeSegmentsDefault);
+
+        $excludeSegmentsDefault = array_map(static function (array $excludeSegment) {
+            return $excludeSegment['provider'] . '::' . $excludeSegment['code'];
+        }, $jobSegmentsManager->getExcludeSegments());
+
+        $form->addMultiSelect('exclude_segment_codes', 'Exclude segments', $segments)
+            ->setHtmlAttribute('class', 'selectpicker')
+            ->setHtmlAttribute('data-live-search', 'true')
+            ->setHtmlAttribute('data-live-search-normalize', 'true')
+            ->setDefaultValue($excludeSegmentsDefault);
 
         $form->addText('context', 'Context')
             ->setNullable()
@@ -86,11 +102,18 @@ class JobFormFactory
     {
         $job = $this->jobsRepository->find($values['job_id']);
 
-        $segment = explode('::', $values['segment_code']);
+        $jobSegmentsManager = new JobSegmentsManager();
+        foreach ($values['include_segment_codes'] as $includeSegment) {
+            [$provider, $code] = explode('::', $includeSegment);
+            $jobSegmentsManager->includeSegment($code, $provider);
+        }
+        foreach ($values['exclude_segment_codes'] as $excludeSegment) {
+            [$provider, $code] = explode('::', $excludeSegment);
+            $jobSegmentsManager->excludeSegment($code, $provider);
+        }
 
         $jobNewData = [
-            'segment_provider' => $segment[0],
-            'segment_code' => $segment[1],
+            'segments' => $jobSegmentsManager->toJson(),
             'context' => $values->context,
         ];