+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Fixed possible performance issue when sending emails. remp/remp#1242
   - The check executed in the `mail:worker` command didn't perform well under certain DB settings and caused unnecessary hold-ups.
 - Fixed `worker:mail` healthcheck not correctly working if worker was occupied with big batch. remp/remp#1240
+- Added support for include and exclude segments in mail jobs. Now you can select multiple include and exclude segments for mail job. remp/remp#1216
 
 ## [1.2.0] - 2023-02-23